refactor(common): consolidate import helpers into resource_import.go
Why this change was needed:
The import helper functions lived in a separate file but were only used
in one place, causing unnecessary file fragmentation and making the code
harder to navigate.

What changed:
- Deleted import_resource_helpers.go and merged its functions into resource_import.go
- Deleted import_resource_helpers_test.go and merged tests into resource_import_test.go
- Renamed resource_slz.go to resource_import.go for clarity
- Renamed resource_slz_import_test.go to resource_import_test.go

Problem solved:
Import-related code is now consolidated in a single file with a clear name,
reducing cognitive load when navigating the codebase.
